requirements.txt
aiohttp-jinja2==1.5.1
aiohttp==3.10.8
aiohttp_session==2.12.0
aiohttp-security==0.4.0
aiohttp-apispec==3.0.0b2
jinja2==3.1.6
pyyaml==6.0.1
cryptography==44.0.1
websockets==15.0
Sphinx==7.1.2
sphinx_rtd_theme==1.3.0
myst-parser==2.0.0
marshmallow==3.20.1
dirhash==0.2.1
marshmallow-enum==1.5.1
ldap3==2.9.1
lxml~=4.9.1  # debrief
reportlab==4.0.4  # debrief
rich==13.7.0
svglib==1.5.1  # debrief
Markdown==3.4.4  # training
dnspython==2.4.2
asyncssh==2.14.1
aioftp~=0.20.0
packaging==23.2
croniter~=3.0.3
setuptools==75.6.0
psutil==5.9.8  # for health monitoring

# Database dependencies
SQLAlchemy==2.0.27
alembic==1.13.1
asyncpg==0.29.0  # PostgreSQL async driver
aiosqlite==0.19.0  # SQLite async driver
greenlet==3.0.3  # Required for SQLAlchemy asyncio support
